AppointmentDTOs added
Altered also respected services and controllers

+++ src/main/java/com/academy/services/AppointmentService.java
@@ -1,48 +1,82 @@
 // AppointmentService.java
 package com.academy.services;
 
+import com.academy.dtos.appointment.AppointmentMapper;
+import com.academy.dtos.appointment.AppointmentRequestDTO;
+import com.academy.dtos.appointment.AppointmentResponseDTO;
 import com.academy.exceptions.EntityNotFoundException;
 import com.academy.models.Appointment;
+import com.academy.models.Member;
 import com.academy.models.service_provider.ServiceProvider;
 import com.academy.repositories.AppointmentRepository;
+import com.academy.repositories.MemberRepository;
+import com.academy.repositories.ServiceProviderRepository;
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
 import java.util.Optional;
+import java.util.stream.Collectors;
 
 @Service
 public class AppointmentService {
 
     private final AppointmentRepository appointmentRepository;
+    private final ServiceProviderRepository serviceProviderRepository;
+    private final AppointmentMapper appointmentMapper;
+    private final MemberRepository memberRepository;
 
-    public AppointmentService(AppointmentRepository appointmentRepository) {
+    @Autowired
+    public AppointmentService(AppointmentRepository appointmentRepository, ServiceProviderRepository serviceProviderRepository, AppointmentMapper appointmentMapper, MemberRepository memberRepository) {
         this.appointmentRepository = appointmentRepository;
+        this.serviceProviderRepository = serviceProviderRepository;
+        this.appointmentMapper = appointmentMapper;
+        this.memberRepository = memberRepository;
     }
 
-    public List<Appointment> getAllAppointments() {
-        return appointmentRepository.findAll();
+    public List<AppointmentResponseDTO> getAllAppointments() {
+        return appointmentRepository.findAll().stream()
+                .map(appointmentMapper::toResponseDTO)
+                .collect(Collectors.toList());
     }
 
-    public Optional<Appointment> getAppointmentById(int id) {
-        return appointmentRepository.findById(id);
+    public Optional<AppointmentResponseDTO> getAppointmentById(int id) {
+        return appointmentRepository.findById(id)
+                .map(appointmentMapper::toResponseDTO);
     }
 
-    public Appointment createAppointment(Appointment appointment) {
-        return appointmentRepository.save(appointment);
+    public AppointmentResponseDTO createAppointment(AppointmentRequestDTO dto) {
+        ServiceProvider serviceProvider = serviceProviderRepository.findById(dto.getServiceProviderId())
+                        .orElseThrow(() -> new EntityNotFoundException(ServiceProvider.class, dto.getServiceProviderId()));
+
+        Appointment appointment = appointmentMapper.toEntity(dto);
+
+        appointment.setServiceProvider(serviceProvider);
+
+        return appointmentMapper.toResponseDTO(appointmentRepository.save(appointment));
     }
 
-    public Appointment updateAppointment(int id, Appointment appointmentDetails) {
+    public AppointmentResponseDTO updateAppointment(int id, AppointmentRequestDTO appointmentDetails) {
         Appointment appointment = appointmentRepository.findById(id)
                 .orElseThrow(() -> new EntityNotFoundException(Appointment.class, id));
-        if(appointmentDetails.getMember() != null) appointment.setMember(appointmentDetails.getMember());
-        if(appointmentDetails.getServiceProvider() != null) appointment.setServiceProvider(appointmentDetails.getServiceProvider());
+        if(appointmentDetails.getMemberId() != null){
+            Member member = memberRepository.findById(appointmentDetails.getMemberId())
+                            .orElseThrow(()-> new EntityNotFoundException(Member.class, appointmentDetails.getMemberId()));
+            appointment.setMember(member);
+        }
+        if(appointmentDetails.getServiceProviderId() != null) {
+            ServiceProvider serviceProvider = serviceProviderRepository.findById(appointmentDetails.getServiceProviderId())
+                    .orElseThrow(()-> new EntityNotFoundException(ServiceProvider.class, appointmentDetails.getServiceProviderId()));
+            appointment.setServiceProvider(serviceProvider);;
+        }
         if(appointmentDetails.getRating() != appointment.getRating() )appointment.setRating(appointmentDetails.getRating());
         if(appointmentDetails.getComment() != null) appointment.setComment(appointmentDetails.getComment());
 
-        return appointmentRepository.save(appointment);
+        return appointmentMapper.toResponseDTO(appointmentRepository.save(appointment));
     }
 
     public void deleteAppointment(int id) {
+        if(!appointmentRepository.existsById(id)) throw new EntityNotFoundException(Appointment.class, id);
         appointmentRepository.deleteById(id);
     }
